Migrate Compose support to be non-NiA-specific
