+++ unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/library/AndroidLibrary.java
@@ -23,7 +23,7 @@ import org.gradle.api.experimental.android.extensions.testing.Testing;
 import org.gradle.api.experimental.android.extensions.CoreLibraryDesugaring;
 import org.gradle.api.experimental.android.extensions.Hilt;
 import org.gradle.api.experimental.android.extensions.KotlinSerialization;
-import org.gradle.api.experimental.android.nia.Compose;
+import org.gradle.api.experimental.android.extensions.Compose;
 import org.gradle.api.experimental.android.nia.Feature;
 import org.gradle.api.provider.Property;
 import org.gradle.api.tasks.Nested;
