--- unified-prototype/unified-plugin/plugin-android/src/main/java/org/gradle/api/experimental/android/nia/Compose.java
@@ -1,12 +1,0 @@
-package org.gradle.api.experimental.android.nia;
-
-import org.gradle.api.provider.Property;
-import org.gradle.declarative.dsl.model.annotations.Restricted;
-
-@Restricted
-public interface Compose {
-    /**
-     * Internal property purposely not exposed to the DSL.
-     */
-    Property<Boolean> getEnabled();
-}
